Prototype 5 - Update for Early Access Version

+++ Assets/Scripts/ManaBar.cs
@@ -7,13 +7,13 @@ public class ManaBar : MonoBehaviour
 {
     [SerializeField] private Slider slider;
 
-    public void SetMaxMana(int mana)
+    public void SetMaxMana(float mana)
     {
         slider.maxValue = mana;
         slider.value = mana;
     }
 
-    public void SetMana(int mana)
+    public void SetMana(float mana)
     {
         slider.value = mana;
     }
